Solution

+++ services/movie_session.py
@@ -1,3 +1,5 @@
+from typing import List
+
 from django.db.models import QuerySet
 
 from db.models import MovieSession, Ticket
@@ -43,9 +45,13 @@ def update_movie_session(
 
 
 def delete_movie_session_by_id(session_id: int) -> None:
-    MovieSession.objects.get(id=session_id).delete()
+    if session_id:
+        MovieSession.objects.get(id=session_id).delete()
+
+    if not session_id:
+        raise f"The id: {session_id} Does Not Exist"
 
 
-def get_taken_seats(movie_session_id: int) -> list:
+def get_taken_seats(movie_session_id: int) -> List[dict[str, int]]:
     tickets = Ticket.objects.filter(movie_session_id=movie_session_id)
     return [{"row": ticket.row, "seat": ticket.seat} for ticket in tickets]
